refactor: route timeline coordinate math through the shared helper
